pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["berezin <isbushcar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
django-bootstrap4 = "^3.0.1"
python-dotenv = "^0.17.1"
gunicorn = "^20.1.0"
whitenoise = "^5.2.0"
psycopg2-binary = "^2.9.1"
django-filter = "^2.4.0"
rollbar = "^0.16.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-cov = "^2.12.0"
python-dotenv = "^0.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
